# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Maia Learn API — implemented MVP
  version: 0.1.0
  description: Free article/Markdown/MP4 courses, resumable upload, private playback,
    curated home, identity and progress. HLS remains planned. Mutations require application/json
    except the authenticated octet-stream chunk route.
servers:
  - url: https://learn.maiaplatform.org/api/v1
  - url: http://localhost:3000/api/v1
paths:
...
  /courses:
    get:
      summary: Search and paginate published course revisions
      parameters:
        - name: q
          in: query
          schema: {type: string, maxLength: 200}
          description: Literal title/summary search; percent and underscore are not wildcards.
        - name: locale
          in: query
          schema: {type: string, enum: [en, pt-BR, es]}
        - name: level
          in: query
          schema: {type: string, enum: [beginner, intermediate, advanced]}
        - name: category
          in: query
          description: Category slug, combined with the other filters using AND. Unknown slugs return no courses.
          schema: {type: string, minLength: 3, maxLength: 100, pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"}
        - name: accessMode
          in: query
          schema: {type: string, enum: [OPEN_FREE, ENROLLED_FREE]}
        - name: sort
          in: query
          schema: {type: string, enum: [newest, title], default: newest}
        - name: page
          in: query
          schema: {type: integer, minimum: 1, maximum: 1000000, default: 1}
        - name: pageSize
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 100}
        - name: format
          in: query
          schema: {type: string, enum: [array, page], default: array}
      responses:
        '200':
          description: Legacy array by default. With format=page, an object containing items, total, page and pageSize. Total and items use the same database snapshot. Filters combine with AND; empty results have total=0 and items=[]. Sorting has a stable ID tie-breaker. Out-of-range pages have empty items.
        '422':
          description: Invalid pagination, filter, sort or search value
...
  /courses/{id}/trailer:
    get:
      summary: Play the selected revision's course trailer
      description: Current published trailers are public regardless of enrollment mode. Drafts require owner/admin access; historical revisions also allow active enrolled learners. Supports HEAD, byte ranges and private no-store caching. Does not record progress.
      parameters:
        - $ref: '#/components/parameters/courseId'
        - name: revisionId
          in: query
          schema: {type: string, format: uuid}
          description: Defaults to the published revision.
      responses:
        '200':
          description: MP4 stream
        '206':
          description: MP4 byte range
        '403':
          description: Revision access denied
        '404':
          description: Trailer missing or not READY
        '416':
          description: Unsatisfiable range
  /courses/{id}/trailer/poster:
    get:
      summary: Trailer poster with the same revision authorization as playback
      parameters:
        - $ref: '#/components/parameters/courseId'
        - name: revisionId
          in: query
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: JPEG poster (HEAD and Range also supported)
        '403':
          description: Revision access denied
        '404':
          description: Trailer or poster unavailable
...
  /lessons/{id}/video:
    parameters:
      - *id003
    get:
      summary: Stream authorized lesson video
      responses: &id004
        '200':
          description: MP4 or JPEG; Cache-Control private, no-store
        '206':
          description: Single byte range
        '403':
          description: No published preview, ownership or active enrollment for revision
        '404':
          description: Media not READY or absent
        '416':
          description: Unsatisfiable/unsupported Range
      parameters:
        - name: Range
          in: header
          schema:
            type: string
          description: One bytes=start-end or suffix range
    head:
      summary: Same authorization and headers as GET, without body
      responses: *id004
  /lessons/{id}/poster:
    parameters:
      - *id003
    get:
      summary: Stream authorized lesson poster
      responses: &id005
        '200':
          description: MP4 or JPEG; Cache-Control private, no-store
        '206':
          description: Single byte range
        '403':
          description: No published preview, ownership or active enrollment for revision
        '404':
          description: Media not READY or absent
        '416':
          description: Unsatisfiable/unsupported Range
      parameters:
        - name: Range
          in: header
          schema:
            type: string
          description: One bytes=start-end or suffix range
    head:
      summary: Same authorization and headers as GET, without body
      responses: *id005
...
  /lessons/{id}/captions/{language}:
    parameters:
      - $ref: '#/components/parameters/lessonId'
      - name: language
        in: path
        required: true
        schema:
          type: string
          enum:
            - en
            - pt-BR
            - es
    get:
      summary: Authorized WebVTT track for a READY lesson video
      responses:
        '200':
          description: Same lesson/revision access as video; private no-store
          content:
            text/vtt:
              schema:
                type: string
        '403':
          description: Enrollment required or wrong revision
        '404':
          description: Missing video or language track
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-sid
      description: Production HTTPS cookie; local HTTP development uses sid.
  parameters:
    courseId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Course UUID or slug
    lessonId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...